+++ src/d3/object.js
@@ -48,8 +48,10 @@ var d3_object = function (arg) {
    * @returns {this}
    */
   this.draw = function () {
-    m_this._update();
-    s_draw();
+    if (m_this.ready) {
+      m_this._update();
+      s_draw();
+    }
     return m_this;
   };